+++ nuget.config
@@ -5,6 +5,5 @@
     <clear />
     <add key="Azure" value="https://pkgs.dev.azure.com/datadoghq/dd-trace-dotnet/_packaging/Public_Feed/nuget/v3/index.json" />
     <add key="nuget" value="https://api.nuget.org/v3/index.json" />
-    <add key="Package source" value="C:\git\dd-trace-dotnet-2.1\tracer\bin\artifacts\nuget" />
   </packageSources>
 </configuration>
