+++ tracer/src/Datadog.Trace/ClrProfiler/DistributedTracer.cs
@@ -4,176 +4,12 @@
 // </copyright>
 
 using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Threading;
 using Datadog.Trace.DuckTyping;
 using Datadog.Trace.Logging;
 
-#pragma warning disable SA1649 // File name should match first type name
-#pragma warning disable SA1402 // File may only contain a single type
-
 namespace Datadog.Trace.ClrProfiler
 {
-    internal interface IDistributedTracer
-    {
-        SpanContext GetSpanContext();
-
-        void SetSpanContext(SpanContext value);
-
-        void LockSamplingPriority();
-
-        SamplingPriority? TrySetSamplingPriority(SamplingPriority? samplingPriority);
-    }
-
-    internal interface ICommonTracer
-    {
-        void LockSamplingPriority();
-
-        int? TrySetSamplingPriority(int? samplingPriority);
-    }
-
-    internal interface IAutomaticTracer : ICommonTracer
-    {
-        object GetDistributedTrace();
-
-        void SetDistributedTrace(object trace);
-
-        void Register(object manualTracer);
-    }
-
-    internal class AutomaticTracer : IAutomaticTracer, IDistributedTracer
-    {
-        private static readonly AsyncLocal<IReadOnlyDictionary<string, string>> DistributedTrace = new();
-        private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor(typeof(AutomaticTracer));
-
-        private ICommonTracer _child;
-
-        private bool _multipleTracers;
-
-        SpanContext IDistributedTracer.GetSpanContext()
-        {
-            if (_multipleTracers)
-            {
-                return SpanContextPropagator.Instance.Extract(DistributedTrace.Value);
-            }
-
-            return null;
-        }
-
-        void IDistributedTracer.SetSpanContext(SpanContext value)
-        {
-            // Locally setting the SpanContext, no need to do anything
-        }
-
-        void IDistributedTracer.LockSamplingPriority()
-        {
-            Log.Information("Automatic - LockSamplingPriority - " + (_child != null).ToString());
-            _child?.LockSamplingPriority();
-        }
-
-        SamplingPriority? IDistributedTracer.TrySetSamplingPriority(SamplingPriority? samplingPriority)
-        {
-            if (_child == null)
-            {
-                Log.Information("Automatic - TrySetSamplingPriority - no child");
-                return samplingPriority;
-            }
-
-            var priority = (SamplingPriority?)_child.TrySetSamplingPriority((int?)samplingPriority);
-
-            Log.Information("Automatic - TrySetSamplingPriority - initial: {initial}, actual: {actual}", samplingPriority, priority);
-
-            return priority;
-        }
-
-        /// <summary>
-        /// Gets the internal distributed trace object
-        /// </summary>
-        /// <returns>Shared distributed trace object instance</returns>
-        public object GetDistributedTrace()
-        {
-            return Tracer.Instance.ActiveScope?.Span.Context;
-        }
-
-        /// <summary>
-        /// Sets the internal distributed trace object
-        /// </summary>
-        /// <param name="value">Shared distributed trace object instance</param>
-        public void SetDistributedTrace(object value)
-        {
-            _multipleTracers = true;
-            DistributedTrace.Value = value as IReadOnlyDictionary<string, string>;
-        }
-
-        public void LockSamplingPriority()
-        {
-            DistributedTracer.LockSamplingPriorityImpl();
-        }
-
-        public int? TrySetSamplingPriority(int? samplingPriority)
-        {
-            return DistributedTracer.TrySetSamplingPriorityImpl(samplingPriority);
-        }
-
-        public void Register(object manualTracer)
-        {
-            Log.Information("Registering {child} as child tracer", manualTracer.GetType());
-            _child = manualTracer.DuckCast<ICommonTracer>();
-        }
-    }
-
-    internal class ManualTracer : IDistributedTracer, ICommonTracer
-    {
-        private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor(typeof(ManualTracer));
-
-        private readonly IAutomaticTracer _parent;
-
-        internal ManualTracer(IAutomaticTracer parent)
-        {
-            _parent = parent ?? throw new ArgumentNullException(nameof(parent));
-            _parent.Register(this);
-        }
-
-        SpanContext IDistributedTracer.GetSpanContext()
-        {
-            var values = _parent.GetDistributedTrace();
-
-            return SpanContextPropagator.Instance.Extract(values as IReadOnlyDictionary<string, string>);
-        }
-
-        void IDistributedTracer.SetSpanContext(SpanContext value)
-        {
-            _parent.SetDistributedTrace(value);
-        }
-
-        void IDistributedTracer.LockSamplingPriority()
-        {
-            Log.Information("Manual - LockSamplingPriority");
-
-            _parent.LockSamplingPriority();
-        }
-
-        SamplingPriority? IDistributedTracer.TrySetSamplingPriority(SamplingPriority? samplingPriority)
-        {
-            var actualPriority = (SamplingPriority?)_parent.TrySetSamplingPriority((int?)samplingPriority);
-
-            Log.Information("Manual - TrySetSamplingPriority - initial: {initial}, actual: {actual}", samplingPriority, actualPriority);
-
-            return actualPriority;
-        }
-
-        public void LockSamplingPriority()
-        {
-            DistributedTracer.LockSamplingPriorityImpl();
-        }
-
-        public int? TrySetSamplingPriority(int? samplingPriority)
-        {
-            return DistributedTracer.TrySetSamplingPriorityImpl(samplingPriority);
-        }
-    }
-
     /// <summary>
     /// Used to distribute traces across multiple versions of the tracer
     /// </summary>
@@ -198,7 +34,7 @@ namespace Datadog.Trace.ClrProfiler
                 {
                     var parentTracer = parent.DuckCast<IAutomaticTracer>();
 
-                    Log.Information("Building manual tracer, connected to " + parent.GetType().Assembly.ToString());
+                    Log.Information("Building manual tracer, connected to {assembly}", parent.GetType().Assembly);
 
                     Instance = new ManualTracer(parentTracer);
                 }
@@ -219,45 +55,5 @@ namespace Datadog.Trace.ClrProfiler
         [Browsable(false)]
         [EditorBrowsable(EditorBrowsableState.Never)]
         public static object GetDistributedTracer() => Instance;
-
-        internal static void LockSamplingPriorityImpl()
-        {
-            var traceContext = Tracer.Instance.ActiveScope?.Span.Context?.TraceContext;
-
-            // If there is no trace context, when a new span is propagated the sampling priority will automatically be locked
-            // because it will be considered as a distributed trace
-            if (traceContext != null)
-            {
-                traceContext.LockSamplingPriority(notifyDistributedTracer: false);
-            }
-        }
-
-        internal static int? TrySetSamplingPriorityImpl(int? samplingPriority)
-        {
-            var traceContext = Tracer.Instance.ActiveScope?.Span.Context?.TraceContext;
-
-            // If there is no trace context, when a new span is propagated the sampling priority will automatically be locked
-            // because it will be considered as a distributed trace
-            if (traceContext != null)
-            {
-                if (traceContext.IsSamplingPriorityLocked())
-                {
-                    var currentSamplingPriority = traceContext.SamplingPriority;
-
-                    if (currentSamplingPriority != null)
-                    {
-                        return (int)currentSamplingPriority.Value;
-                    }
-
-                    Log.Warning("SamplingPriority is locked without value");
-                }
-                else
-                {
-                    traceContext.SetSamplingPriority((SamplingPriority?)samplingPriority, notifyDistributedTracer: false);
-                }
-            }
-
-            return samplingPriority;
-        }
     }
 }
