+++ tracer/src/Datadog.Trace/ClrProfiler/IAutomaticTracer.cs
@@ -1,0 +1,16 @@
+// <copyright file="IAutomaticTracer.cs" company="Datadog">
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
+// </copyright>
+
+namespace Datadog.Trace.ClrProfiler
+{
+    internal interface IAutomaticTracer : ICommonTracer
+    {
+        object GetDistributedTrace();
+
+        void SetDistributedTrace(object trace);
+
+        void Register(object manualTracer);
+    }
+}
