+++ tracer/test/Datadog.Trace.ClrProfiler.IntegrationTests/AspNet/AspNetVersionConflictTests.cs
@@ -26,7 +26,9 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.AspNet
 
             _iisFixture = iisFixture;
             _iisFixture.ShutdownPath = "/home/shutdown";
-            _iisFixture.TryStartIis(this, IisAppType.AspNetIntegrated);
+            // There is an issue in the TracingHttpModule that causes the parent trace to be locked, making the test useless
+            // This code is not run when IIS is in classic mode, so using that as a workaround
+            _iisFixture.TryStartIis(this, IisAppType.AspNetClassic);
         }
 
         [SkippableFact]
@@ -69,19 +71,19 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.AspNet
             httpSpan.Name.Should().Be("http.request");
         }
 
-        [SkippableFact]
+        [SkippableTheory]
+        [InlineData(true)]
+        [InlineData(false)]
         [Trait("Category", "EndToEnd")]
         [Trait("RunOnWindows", "True")]
         [Trait("LoadFromGAC", "True")]
-        public async Task Sampling()
+        public async Task Sampling(bool parentTrace)
         {
             // 6 spans for the base request: aspnet.request / aspnet-mvc.request / Manual / http.request / http.request / Child
             // + 2 * 2 spans for the outgoing requests: aspnet.request / aspnet-mvc.request
             const int expectedSpans = 10;
 
-            var spans = await GetWebServerSpans("/home/sampling", _iisFixture.Agent, _iisFixture.HttpPort, System.Net.HttpStatusCode.OK, expectedSpans, filterServerSpans: false);
-
-            spans.Should().HaveCount(expectedSpans);
+            var spans = await GetWebServerSpans($"/home/sampling?parentTrace={parentTrace}", _iisFixture.Agent, _iisFixture.HttpPort, System.Net.HttpStatusCode.OK, expectedSpans, filterServerSpans: false);
 
             foreach (var span in spans)
             {
@@ -95,11 +97,30 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.AspNet
                 Output.WriteLine($"{span.Name} - {span.TraceId} - {span.SpanId} - {span.ParentId} - {span.Resource} - {samplingPriority}");
             }
 
-            // Make sure there is only one root span
-            spans.Where(s => s.ParentId == null).Should().HaveCount(1);
+            // Make sure there is no extra root span
+            spans.Where(s => s.ParentId == null).Should().HaveCount(parentTrace ? 1 : 2);
 
             // The sampling priority should be UserKeep for all spans
-            spans.Should().OnlyContain(s => !s.Metrics.ContainsKey(Metrics.SamplingPriority) || s.Metrics[Metrics.SamplingPriority] == 2);
+            spans.Should().OnlyContain(s => VerifySpan(s, parentTrace));
+        }
+
+        private static bool VerifySpan(MockTracerAgent.Span span, bool parentTrace)
+        {
+            if (!span.Metrics.ContainsKey(Metrics.SamplingPriority))
+            {
+                return true;
+            }
+
+            if (!parentTrace)
+            {
+                // The root asp.net trace has an automatic priority
+                if (span.Name == "aspnet.request" && span.Resource == "GET /home/sampling")
+                {
+                    return span.Metrics[Metrics.SamplingPriority] == 1;
+                }
+            }
+
+            return span.Metrics[Metrics.SamplingPriority] == 2;
         }
     }
 }
