--- tracer/test/test-applications/aspnet/Samples.AspNet.VersionConflict/Areas/Datadog/Views/DogHouse/Index.cshtml
@@ -1,6 +1,0 @@
-﻿
-@{
-    ViewBag.Title = "Datadog";
-}
-
-<h2>Datadog Area Home</h2>
