--- tracer/test/test-applications/aspnet/Samples.AspNet.VersionConflict/Areas/Datadog/Views/Shared/_Layout.cshtml
@@ -1,17 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>@ViewBag.Title</title>
-</head>
-<body>
-    <div class="container body-content">
-        @RenderBody()
-        <hr />
-        <footer>
-            <p>&copy; @DateTime.Now.Year</p>
-        </footer>
-    </div>
-</body>
-</html>
