--- tracer/test/test-applications/aspnet/Samples.AspNet.VersionConflict/Areas/Datadog/Views/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "~/Areas/Datadog/Views/Shared/_Layout.cshtml";
-}
