+++ tracer/test/test-applications/aspnet/Samples.AspNet.VersionConflict/Global.asax.cs
@@ -9,7 +9,6 @@ namespace Samples.AspNet.VersionConflict
     {
         protected void Application_Start()
         {
-            GlobalConfiguration.Configure(WebApiConfig.Register);
             AreaRegistration.RegisterAllAreas();
             RouteConfig.RegisterRoutes(RouteTable.Routes);
         }
