--- tracer/test/test-applications/aspnet/Samples.AspNet.VersionConflict/nuget.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <packageSources>
-    <add key="local" value="C:\git\dd-trace-dotnet-2.1\tracer\bin\artifacts\nuget" />
-    <!--<add key="Azure_Devops" value="https://pkgs.dev.azure.com/datadoghq/dd-trace-dotnet/_packaging/Public_Feed/nuget/v3/index.json" />-->
-  </packageSources>
-</configuration>
